♻️ :: add CategoryNoticeTableView Extension
